fix(ui): normalize scores to 0–1 scale + Bar max prop
Score thresholds changed from /100 to /1.0 throughout dashboard,
rubric pages, runs, and reports. Bar component accepts max prop.
scoreTone thresholds adjusted (75 → 0.75). Mock data aligned.

+++ app/src/app/rubrics/actions.ts
@@ -27,7 +27,7 @@ function parseDimensions(formData: FormData): Dimension[] {
       name: (formData.get(`dim_name_${i}`) as string).trim(),
       method: (formData.get(`dim_method_${i}`) as string).trim(),
       weight: parseFloat(formData.get(`dim_weight_${i}`) as string) || 0,
-      threshold: parseInt(formData.get(`dim_threshold_${i}`) as string) || 0,
+      threshold: parseFloat(formData.get(`dim_threshold_${i}`) as string) || 0,
     });
     i++;
   }
